*sighhh*

+++ strace/_strace.h
@@ -11,6 +11,6 @@
 #include "syscalls.h"
 
 int parent_func(pid_t);
-const char *get_syscall_name(int syscall_num);
+const char *get_syscall_name(size_t syscall_num);
 
 #endif
